+++ demos/browser/app/meetingV2/meetingV2.ts
@@ -173,9 +173,10 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
       this.region = (document.getElementById('inputRegion') as HTMLInputElement).value;
       new AsyncScheduler().start(
         async (): Promise<void> => {
+          let chimeMeetingId: string = '';
           this.showProgress('progress-authenticate');
           try {
-            await this.authenticate();
+            chimeMeetingId = await this.authenticate();
           } catch (error) {
             (document.getElementById(
               'failed-meeting'
@@ -188,6 +189,9 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
           (document.getElementById(
             'meeting-id'
           ) as HTMLSpanElement).innerHTML = `${this.meeting} (${this.region})`;
+          (document.getElementById(
+            'chime-meeting-id'
+          ) as HTMLSpanElement).innerHTML = `${chimeMeetingId}`;
           (document.getElementById('info-meeting') as HTMLSpanElement).innerHTML = this.meeting;
           (document.getElementById('info-name') as HTMLSpanElement).innerHTML = this.name;
           this.switchToFlow('flow-devices');
@@ -1074,7 +1078,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
     }
   }
 
-  async authenticate(): Promise<void> {
+  async authenticate(): Promise<string> {
     let joinInfo = (await this.joinMeeting()).JoinInfo;
     await this.initializeMeetingSession(
       new MeetingSessionConfiguration(joinInfo.Meeting, joinInfo.Attendee)
@@ -1082,6 +1086,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
     const url = new URL(window.location.href);
     url.searchParams.set('m', this.meeting);
     history.replaceState({}, `${this.meeting}`, url.toString());
+    return joinInfo.Meeting.MeetingId;
   }
 
   log(str: string): void {